requirements.txt
absl-py==1.3.0
cachetools==5.2.0
certifi==2022.9.24
click==8.1.3
cycler==0.11.0
Flask==2.2.2
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
importlib-metadata==5.0.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-cpu==2.8.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.10.0
astunparse==1.6.3
brotlipy==0.7.0
contourpy==1.0.5
flatbuffers==22.10.26
fonttools==4.38.0
itsdangerous==2.1.2
libclang==14.0.6
packaging==21.3
pycosat==0.6.3
threadpoolctl==3.1.0
typing_extensions==4.4.0
